+++ SchemeCreator.csproj
@@ -136,9 +136,9 @@
     <Compile Include="Data\Interfaces\IValueHolder.cs" />
     <Compile Include="Data\Models\WireConnection.cs" />
     <Compile Include="Data\Models\Enums\ConnectionTypeEnum.cs" />
+    <Compile Include="Data\Services\Alignment\Aliner.cs" />
     <Compile Include="Data\Services\Colorer.cs" />
     <Compile Include="Data\Services\GateWorkPatterns.cs" />
-    <Compile Include="Data\Services\Aliner.cs" />
     <Compile Include="Data\Services\MessageData.cs" />
     <Compile Include="Data\Services\Messages.cs" />
     <Compile Include="Data\Services\Navigation\SchemeNavigationService.cs" />
